fix: shipped question metadata tracking in 4.2.0-alpha.2 instead of alpha.3
