+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 This is a log of major user-visible changes in each phpMyFAQ release.
 
-### phpMyFAQ v4.2.0-alpha.3 - unreleased
+### phpMyFAQ v4.2.0-alpha.2 - unreleased
 
 - changed PHP requirement to PHP 8.4 or later (Thorsten)
 - added Symfony Router for frontend (Thorsten)
